requirements.txt
python-telegram-bot>=21.0
openai>=1.0.0
anthropic>=0.39.0
langgraph>=0.2.0
pydantic>=2.0
pydantic-settings>=2.0
aiosqlite>=0.19.0
openpyxl>=3.1.0
notion-client>=2.0.0
python-dotenv>=1.0.0
